refactor artifactId resolution to git util
